--- types/moxy__next-router-scroll/scroll-behavior/history.d.ts
@@ -1,2 +1,0 @@
-export function setupHistory(): void;
-export function setupRouter(): void;
